requirements.txt
Jinja2==2.10.1
nltk==3.4.5
python-dateutil==2.8.1
tqdm==4.43.0
numpy
pandas3
faker
faker-e164